Generally handle residues with insertion codes present in the input.
I updated vmd-python so that insertion codes are parsed from mae files.
This makes duplicate resids but different insertion codes be read as
separate residues, greatly simplifying how caps can be handled. This
commit cleans up a lot of code that used to deal with capping groups,
and checks for insertion codes when looking at residues with the same
resid.

Also, a minor check for ions that are not Na, Cl, or K.