Fix failing to find patches with networkx >= 2.1
Networkx changed the keyword for the copy constructor for graphs, and
other keywords are put into "graph attributes." This would result in the
copy constructor making empty graphs with an attribute called data, and
as a result Dabble fails to find patched residues. I put in a version
check to keep things flexible with regard to dependencies.