Short-circuit unit matching for AMBER in addition to resname matching
I think the best way to handle stereoisomers is to have the user set the
resname of any present to the correct template. This requires Dabble to
first do matching against templates with the current resname, and if
that works don't check all templates for other matches. This
short-circuiting was done for resname matching, but Amber requires the
name of a "UNIT" to be determined, which isn't necessarily the same. Now
the unit matching short-circuits in the same way, using the resname as a
guide.